+++ libvmaf/src/feature/funque_filters.c
@@ -27,20 +27,20 @@
 #include "offset.h"
 #include "funque_filters.h"
 
-void funque_dwt2(float *src, dwt2buffers *dwt2_dst, ptrdiff_t dst_stride, int width, int height)
+void funque_dwt2(double *src, dwt2buffers *dwt2_dst, ptrdiff_t dst_stride, int width, int height)
 {
-    int dst_px_stride = dst_stride / sizeof(float);
-    float filter_coeff_lo[2] = {0.707106781,  0.707106781};
-    float filter_coeff_hi[2] = {0.707106781, -0.707106781};
+    int dst_px_stride = dst_stride / sizeof(double);
+    double filter_coeff_lo[2] = {0.707106781,  0.707106781};
+    double filter_coeff_hi[2] = {0.707106781, -0.707106781};
 
-    float *tmplo = aligned_malloc(ALIGN_CEIL(width * sizeof(float)), MAX_ALIGN);
-    float *tmphi = aligned_malloc(ALIGN_CEIL(width * sizeof(float)), MAX_ALIGN);
+    double *tmplo = aligned_malloc(ALIGN_CEIL(width * sizeof(double)), MAX_ALIGN);
+    double *tmphi = aligned_malloc(ALIGN_CEIL(width * sizeof(double)), MAX_ALIGN);
 
-    float *band_a = dwt2_dst->bands[0];
-    float *band_h = dwt2_dst->bands[1];
-    float *band_v = dwt2_dst->bands[2];
-    float *band_d = dwt2_dst->bands[3];
-    float accum;
+    double *band_a = dwt2_dst->bands[0];
+    double *band_h = dwt2_dst->bands[1];
+    double *band_v = dwt2_dst->bands[2];
+    double *band_d = dwt2_dst->bands[3];
+    double accum;
     int16_t row_idx0, row_idx1, col_idx0, col_idx1;
     for (unsigned i=0; i < (height+1)/2; ++i)
     {
@@ -95,10 +95,10 @@ void funque_dwt2(float *src, dwt2buffers *dwt2_dst, ptrdiff_t dst_stride, int wi
 }
 
 //Convolution using coefficients from python workspace
-void spatial_filter(float *src, float *dst, ptrdiff_t dst_stride, int width, int height)
+void spatial_filter(double *src, double *dst, ptrdiff_t dst_stride, int width, int height)
 {
     //Copied the coefficients from python coefficients
-    float filter_coeffs[21] = {-0.01373464, -0.01608515, -0.01890698, -0.02215702, -0.02546262, 
+    double filter_coeffs[21] = {-0.01373464, -0.01608515, -0.01890698, -0.02215702, -0.02546262, 
                              -0.02742965, -0.02361034, -0.00100996,  0.07137023,  0.22121922,
                              0.3279824 ,  0.22121922,  0.07137023, -0.00100996, -0.02361034,
                              -0.02742965, -0.02546262, -0.02215702, -0.01890698, -0.01608515,
@@ -107,8 +107,8 @@ void spatial_filter(float *src, float *dst, ptrdiff_t dst_stride, int width, int
     int src_px_stride = width;
     int dst_px_stride = width;
 
-    float *tmp = aligned_malloc(ALIGN_CEIL(src_px_stride * sizeof(float)), MAX_ALIGN);
-    float fcoeff, imgcoeff;
+    double *tmp = aligned_malloc(ALIGN_CEIL(src_px_stride * sizeof(double)), MAX_ALIGN);
+    double fcoeff, imgcoeff;
 
     int i, j, fi, fj, ii, jj;
     int fwidth = 21;
@@ -134,7 +134,7 @@ void spatial_filter(float *src, float *dst, ptrdiff_t dst_stride, int width, int
 
         /* Horizontal pass. */
         for (j = 0; j < width; ++j) {
-            float accum = 0;
+            double accum = 0;
 
             for (fj = 0; fj < fwidth; ++fj) {
                 fcoeff = filter_coeffs[fj];
@@ -156,14 +156,14 @@ void spatial_filter(float *src, float *dst, ptrdiff_t dst_stride, int width, int
     return;
 }
 
-void normalize_bitdepth(float *src, float *dst, int scaler, ptrdiff_t dst_stride, int width, int height)
+void normalize_bitdepth(double *src, double *dst, int scaler, ptrdiff_t dst_stride, int width, int height)
 {
     for (unsigned i = 0; i < height; i++) {
         for (unsigned j = 0; j < width; j++) {
             dst[j] = src[j] / scaler;
         }
-        dst += dst_stride / sizeof(float);
-        src += dst_stride / sizeof(float);
+        dst += dst_stride / sizeof(double);
+        src += dst_stride / sizeof(double);
     }
     return;
 }
