+++ libvmaf/src/feature/funque_ssim.h
@@ -16,4 +16,4 @@
  *
  */
 
-int compute_ssim_funque(dwt2buffers *ref, dwt2buffers *dist, double *score, int max_val, float K1, float K2);
+int compute_ssim_funque(dwt2buffers *ref, dwt2buffers *dist, double *score, int max_val, float K1, double K2);
